Board displays and isValidPiece method tests to make sure user input is valid
